+++ components/dm/command/patch.go
@@ -13,24 +13,8 @@
 
 package command
 
-/*
 import (
-	"errors"
-	"fmt"
-	"os"
-	"os/exec"
-	"path"
-
-	"github.com/joomcode/errorx"
 	perrs "github.com/pingcap/errors"
-	"github.com/pingcap/tiup/pkg/cluster/clusterutil"
-	"github.com/pingcap/tiup/pkg/cluster/meta"
-	operator "github.com/pingcap/tiup/pkg/cluster/operation"
-	"github.com/pingcap/tiup/pkg/cluster/task"
-	tiupmeta "github.com/pingcap/tiup/pkg/environment"
-	"github.com/pingcap/tiup/pkg/repository/v0manifest"
-	"github.com/pingcap/tiup/pkg/set"
-	tiuputils "github.com/pingcap/tiup/pkg/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -53,7 +37,10 @@ func newPatchCmd() *cobra.Command {
 			if len(gOpt.Nodes) == 0 && len(gOpt.Roles) == 0 {
 				return perrs.New("the flag -R or -N must be specified at least one")
 			}
-			return patch(args[0], args[1], gOpt, overwrite)
+
+			clusterName := args[0]
+
+			return deployer.Patch(clusterName, args[1], gOpt, overwrite)
 		},
 	}
 
@@ -63,136 +50,3 @@ func newPatchCmd() *cobra.Command {
 	cmd.Flags().Int64Var(&gOpt.APITimeout, "transfer-timeout", 300, "Timeout in seconds when transferring dm-master leaders")
 	return cmd
 }
-
-func patch(clusterName, packagePath string, options operator.Options, overwrite bool) error {
-	if tiuputils.IsNotExist(meta.ClusterPath(clusterName, meta.MetaFileName)) {
-		return perrs.Errorf("cannot patch non-exists cluster %s", clusterName)
-	}
-
-	if exist := tiuputils.IsExist(packagePath); !exist {
-		return perrs.New("specified package not exists")
-	}
-
-	metadata, err := meta.DMMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ValidateErr) {
-		return err
-	}
-
-	insts, err := instancesToPatch(metadata, options)
-	if err != nil {
-		return err
-	}
-	if err := checkPackage(clusterName, insts[0].ComponentName(), packagePath); err != nil {
-		return err
-	}
-
-	var replacePackageTasks []task.Task
-	for _, inst := range insts {
-		deployDir := clusterutil.Abs(metadata.User, inst.DeployDir())
-		tb := task.NewBuilder()
-		tb.BackupComponent(inst.ComponentName(), metadata.Version, inst.GetHost(), deployDir).
-			InstallPackage(packagePath, inst.GetHost(), deployDir)
-		replacePackageTasks = append(replacePackageTasks, tb.Build())
-	}
-
-	t := task.NewBuilder().
-		SSHKeySet(
-			meta.ClusterPath(clusterName, "ssh", "id_rsa"),
-			meta.ClusterPath(clusterName, "ssh", "id_rsa.pub")).
-		ClusterSSH(metadata.Topology, metadata.User, gOpt.SSHTimeout).
-		Parallel(replacePackageTasks...).
-		ClusterOperate(metadata.Topology, operator.UpgradeOperation, options).
-		Build()
-
-	if err := t.Execute(task.NewContext()); err != nil {
-		if errorx.Cast(err) != nil {
-			// FIXME: Map possible task errors and give suggestions.
-			return err
-		}
-		return perrs.Trace(err)
-	}
-
-	if overwrite {
-		if err := overwritePatch(clusterName, insts[0].ComponentName(), packagePath); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func instancesToPatch(metadata *meta.DMMeta, options operator.Options) ([]meta.Instance, error) {
-	roleFilter := set.NewStringSet(options.Roles...)
-	nodeFilter := set.NewStringSet(options.Nodes...)
-	components := metadata.Topology.ComponentsByStartOrder()
-	components = operator.FilterComponent(components, roleFilter)
-
-	instances := []meta.Instance{}
-	comps := []string{}
-	for _, com := range components {
-		insts := operator.FilterInstance(com.Instances(), nodeFilter)
-		if len(insts) > 0 {
-			comps = append(comps, com.Name())
-		}
-		instances = append(instances, insts...)
-	}
-	if len(comps) > 1 {
-		return nil, fmt.Errorf("can't patch more than one component at once: %v", comps)
-	}
-
-	if len(instances) == 0 {
-		return nil, fmt.Errorf("no instance found on specifid role(%v) and nodes(%v)", options.Roles, options.Nodes)
-	}
-
-	return instances, nil
-}
-
-func checkPackage(clusterName, comp, packagePath string) error {
-	metadata, err := meta.DMMetadata(clusterName)
-	if err != nil && !errors.Is(perrs.Cause(err), meta.ValidateErr) {
-		return err
-	}
-	manifest, err := tiupmeta.GlobalEnv().Repository().ComponentVersions(comp)
-	if err != nil {
-		return err
-	}
-	ver := meta.ComponentVersion(comp, metadata.Version)
-	versionInfo, found := manifest.FindVersion(v0manifest.Version(ver))
-	if !found {
-		return fmt.Errorf("cannot found version %v in %s manifest", ver, comp)
-	}
-
-	checksum, err := tiuputils.Checksum(packagePath)
-	if err != nil {
-		return err
-	}
-	cacheDir := meta.ClusterPath(clusterName, "cache", comp+"-"+checksum[:7])
-	if err := os.MkdirAll(cacheDir, 0755); err != nil {
-		return err
-	}
-	if err := exec.Command("tar", "-xvf", packagePath, "-C", cacheDir).Run(); err != nil {
-		return err
-	}
-
-	if exists := tiuputils.IsExist(path.Join(cacheDir, versionInfo.Entry)); !exists {
-		return fmt.Errorf("entry %s not found in package %s", versionInfo.Entry, packagePath)
-	}
-
-	return nil
-}
-
-func overwritePatch(clusterName, comp, packagePath string) error {
-	if err := os.MkdirAll(meta.ClusterPath(clusterName, meta.PatchDirName), 0755); err != nil {
-		return err
-	}
-	checksum, err := tiuputils.Checksum(packagePath)
-	if err != nil {
-		return err
-	}
-	tg := meta.ClusterPath(clusterName, meta.PatchDirName, comp+"-"+checksum[:7]+".tar.gz")
-	if err := tiuputils.CopyFile(packagePath, tg); err != nil {
-		return err
-	}
-	return os.Symlink(tg, meta.ClusterPath(clusterName, meta.PatchDirName, comp+".tar.gz"))
-}
-*/
