+++ pkg/cluster/deploy/deploy.go
@@ -7,6 +7,8 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
+	"os/exec"
+	"path"
 	"sort"
 	"strings"
 
@@ -669,6 +671,65 @@ func (d *Deployer) Upgrade(clusterName string, clusterVersion string, opt operat
 	return nil
 }
 
+// Patch the cluster.
+func (d *Deployer) Patch(clusterName string, packagePath string, opt operator.Options, overwrite bool) error {
+	metadata, err := d.meta(clusterName)
+	if err != nil {
+		return perrs.AddStack(err)
+	}
+
+	topo := metadata.GetTopology()
+	base := metadata.GetBaseMeta()
+
+	if exist := utils.IsExist(packagePath); !exist {
+		return perrs.New("specified package not exists")
+	}
+
+	insts, err := instancesToPatch(topo, opt)
+	if err != nil {
+		return err
+	}
+	if err := checkPackage(clusterName, insts[0].ComponentName(), insts[0].OS(), insts[0].Arch(), packagePath); err != nil {
+		return err
+	}
+
+	var replacePackageTasks []task.Task
+	for _, inst := range insts {
+		deployDir := clusterutil.Abs(base.User, inst.DeployDir())
+		tb := task.NewBuilder()
+		tb.BackupComponent(inst.ComponentName(), base.Version, inst.GetHost(), deployDir).
+			InstallPackage(packagePath, inst.GetHost(), deployDir)
+		replacePackageTasks = append(replacePackageTasks, tb.Build())
+	}
+
+	t := task.NewBuilder().
+		SSHKeySet(
+			spec.ClusterPath(clusterName, "ssh", "id_rsa"),
+			spec.ClusterPath(clusterName, "ssh", "id_rsa.pub")).
+		ClusterSSH(topo, base.User, opt.SSHTimeout).
+		Parallel(replacePackageTasks...).
+		Serial(task.NewFunc("UpgradeCluster", func(ctx *task.Context) error {
+			return operator.Upgrade(ctx, topo, opt)
+		})).
+		Build()
+
+	if err := t.Execute(task.NewContext()); err != nil {
+		if errorx.Cast(err) != nil {
+			// FIXME: Map possible task errors and give suggestions.
+			return err
+		}
+		return perrs.Trace(err)
+	}
+
+	if overwrite {
+		if err := overwritePatch(clusterName, insts[0].ComponentName(), packagePath); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 func (d *Deployer) meta(name string) (metadata spec.Metadata, err error) {
 	exist, err := d.specManager.Exist(name)
 	if err != nil {
@@ -814,4 +875,89 @@ func versionCompare(curVersion, newVersion string) error {
 type componentInfo struct {
 	component string
 	version   string
+}
+
+func instancesToPatch(topo spec.Topology, options operator.Options) ([]spec.Instance, error) {
+	roleFilter := set.NewStringSet(options.Roles...)
+	nodeFilter := set.NewStringSet(options.Nodes...)
+	components := topo.ComponentsByStartOrder()
+	components = operator.FilterComponent(components, roleFilter)
+
+	instances := []spec.Instance{}
+	comps := []string{}
+	for _, com := range components {
+		insts := operator.FilterInstance(com.Instances(), nodeFilter)
+		if len(insts) > 0 {
+			comps = append(comps, com.Name())
+		}
+		instances = append(instances, insts...)
+	}
+	if len(comps) > 1 {
+		return nil, fmt.Errorf("can't patch more than one component at once: %v", comps)
+	}
+
+	if len(instances) == 0 {
+		return nil, fmt.Errorf("no instance found on specifid role(%v) and nodes(%v)", options.Roles, options.Nodes)
+	}
+
+	return instances, nil
+}
+
+func checkPackage(clusterName, comp, nodeOS, arch, packagePath string) error {
+	metadata, err := spec.ClusterMetadata(clusterName)
+	if err != nil && !errors.Is(perrs.Cause(err), meta.ErrValidate) {
+		return err
+	}
+
+	ver := spec.ComponentVersion(comp, metadata.Version)
+	repo, err := clusterutil.NewRepository(nodeOS, arch)
+	if err != nil {
+		return err
+	}
+	entry, err := repo.ComponentBinEntry(comp, ver)
+	if err != nil {
+		return err
+	}
+
+	checksum, err := utils.Checksum(packagePath)
+	if err != nil {
+		return err
+	}
+	cacheDir := spec.ClusterPath(clusterName, "cache", comp+"-"+checksum[:7])
+	if err := os.MkdirAll(cacheDir, 0755); err != nil {
+		return err
+	}
+	if err := exec.Command("tar", "-xvf", packagePath, "-C", cacheDir).Run(); err != nil {
+		return err
+	}
+
+	if exists := utils.IsExist(path.Join(cacheDir, entry)); !exists {
+		return fmt.Errorf("entry %s not found in package %s", entry, packagePath)
+	}
+
+	return nil
+}
+
+func overwritePatch(clusterName, comp, packagePath string) error {
+	if err := os.MkdirAll(spec.ClusterPath(clusterName, spec.PatchDirName), 0755); err != nil {
+		return err
+	}
+
+	checksum, err := utils.Checksum(packagePath)
+	if err != nil {
+		return err
+	}
+
+	tg := spec.ClusterPath(clusterName, spec.PatchDirName, comp+"-"+checksum[:7]+".tar.gz")
+	if !utils.IsExist(tg) {
+		if err := utils.CopyFile(packagePath, tg); err != nil {
+			return err
+		}
+	}
+
+	symlink := spec.ClusterPath(clusterName, spec.PatchDirName, comp+".tar.gz")
+	if utils.IsSymExist(symlink) {
+		os.Remove(symlink)
+	}
+	return os.Symlink(tg, symlink)
 }
